Implemented pitch and roll movement
Optimizations to ray tracing routines
Added sphere as a primitive
Added shadows
Added dithering for performance

+++ src/Intersectable.java
@@ -1,8 +1,6 @@
-import java.awt.Color;
-
 public interface Intersectable {
 
-	public static final float K_EPSILON = 0.0001f;
+	public static final double K_EPSILON = 0.0001f;
 	
 	public Intersection getIntersection(Ray3 ray);
 }
